score/pyproject.toml
[project]
name = "score"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.11"
dependencies = [
    "ipykernel>=6.29.5",
    "numpy>=1.26.4",
    "pandas>=2.3.0",
    "sacrebleu>=2.5.1",
    "torch>=2.7.1",
    "unbabel-comet>=2.2.6",
]

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
